Refactor formatters and collectors: introduce helper modules, streamline logger initialization, and reorganize depth formatting
