+++ lib/trace_viz/formatters/base_formatter.rb
@@ -1,16 +1,12 @@
 # frozen_string_literal: true
 
-require_relative "helpers/indent_helper"
-require_relative "helpers/depth_helper"
-require_relative "helpers/time_helper"
-require_relative "helpers/params_helper"
-require_relative "helpers/source_helper"
-require_relative "helpers/result_helper"
-require_relative "helpers/method_details_helper"
+require "trace_viz/helpers"
 
 module TraceViz
   module Formatters
     class BaseFormatter
+      include Helpers::ConfigHelper
+
       def call
         raise NotImplementedError
       end
