+++ lib/trace_viz/formatters/helpers.rb
@@ -1,0 +1,10 @@
+# frozen_string_literal: true
+
+Dir[File.join(__dir__, "helpers/**/*.rb")].each { |file| require_relative file }
+
+module TraceViz
+  module Formatters
+    module Helpers
+    end
+  end
+end
