+++ lib/trace_viz/formatters/log/base_formatter.rb
@@ -1,0 +1,13 @@
+# frozen_string_literal: true
+
+require_relative "../trace_data_formatter"
+
+module TraceViz
+  module Formatters
+    module Log
+      class BaseFormatter < TraceDataFormatter
+        include Helpers::Log::ColorHelper
+      end
+    end
+  end
+end
