+++ lib/trace_viz/formatters/trace_data_formatter.rb
@@ -1,13 +1,7 @@
 # frozen_string_literal: true
 
 require_relative "base_formatter"
-require_relative "helpers/indent_helper"
-require_relative "helpers/depth_helper"
-require_relative "helpers/time_helper"
-require_relative "helpers/params_helper"
-require_relative "helpers/source_helper"
-require_relative "helpers/result_helper"
-require_relative "helpers/method_details_helper"
+require_relative "helpers"
 
 module TraceViz
   module Formatters
@@ -29,12 +23,11 @@ module TraceViz
       def initialize(trace_data)
         super()
         @trace_data = trace_data
-        @config = @trace_data.config
       end
 
       private
 
-      attr_reader :trace_data, :config
+      attr_reader :trace_data
     end
   end
 end
