--- lib/trace_viz/loggers/trace_formatters/helpers/method_name_helper.rb
@@ -1,29 +1,0 @@
-# frozen_string_literal: true
-
-module TraceViz
-  module Loggers
-    module TraceFormatters
-      module Helpers
-        module MethodNameHelper
-          def format_method_name(trace_data, config)
-            return unless config.general[:show_method_name]
-
-            "#{format_class_name(trace_data)}#{format_action_name(trace_data)}"
-          end
-
-          private
-
-          def format_class_name(trace_data)
-            klass = colorize(trace_data.klass, :trace_method_class)
-            method_sign = colorize("#", :trace_method_sign)
-            [klass, method_sign].join
-          end
-
-          def format_action_name(trace_data)
-            colorize(trace_data.action, :trace_method_action)
-          end
-        end
-      end
-    end
-  end
-end
